Move remap parsing to remap module
